Start splitting to common and velocity module
