+++ gradle.properties
@@ -1,0 +1,1 @@
+org.gradle.warning.mode=all
