+++ settings.gradle
@@ -1,2 +1,4 @@
 rootProject.name = 'Paxy'
 
+include 'common', 'velocity'
+
