+++ velocity/build.gradle
@@ -1,0 +1,37 @@
+plugins {
+    id 'com.github.johnrengelman.shadow' version '6.1.0'
+}
+
+def velocity_version = '1.1.4-SNAPSHOT'
+
+repositories {
+    // Velocity Repo
+    maven { url 'https://nexus.velocitypowered.com/repository/maven-public/' }
+}
+
+dependencies {
+    compile project(':common')
+
+    implementation "com.velocitypowered:velocity-api:${velocity_version}"
+    annotationProcessor "com.velocitypowered:velocity-api:${velocity_version}"
+}
+
+shadowJar {
+    archiveBaseName.set('paxy-velocity')
+    archiveVersion.set("$project.version")
+    archiveClassifier.set('')
+
+    dependencies {
+        // doesnt exclude transitive deps
+//        exclude dependency("io.netty:netty-all:.*")
+//        exclude dependency("com.velocitypowered:velocity-api")
+    }
+
+    // Only merge dependencies with the 'compile' configuration
+    //todo doesnt seem to be working?
+    configurations = [project.configurations.compile]
+}
+
+artifacts {
+    archives shadowJar
+}
