+++ velocity/src/main/java/net/minestom/paxy/velocity/handlers/VelocityDecoderHandler.java
@@ -1,4 +1,4 @@
-package net.minestom.paxy.handlers;
+package net.minestom.paxy.velocity.handlers;
 
 import com.github.steveice10.mc.protocol.data.SubProtocol;
 import com.github.steveice10.mc.protocol.packet.login.server.LoginSuccessPacket;
@@ -8,6 +8,8 @@ import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.MessageToByteEncoder;
 import io.netty.handler.codec.MessageToMessageDecoder;
+import net.minestom.paxy.handlers.Handlers;
+import net.minestom.paxy.handlers.PaxyProtocol;
 import net.minestom.paxy.utils.BufUtils;
 import net.minestom.paxy.utils.PipelineUtil;
 import net.minestom.paxy.utils.ProtocolUtils;
